Make Settings UI Window More Compact
Fixing the bug in the "Emulation" tab in the Settings UI has the side effect of increasing the minimum size of the Settings UI window by 48 pixels. It's been brought to my attention that this might make it more difficult and cumbersome to use the UI on a lower-resolution display.

By making some minor changes to the layout of the "Emulation" tab, I was able to reduce the minimum height of the window by 40 pixels. This means that the minimum height for the Settings UI window is now only eight pixels larger than it was before the bugfix was applied.