Add "Try to adjust 16:9 game to fit" option (ultrawidescreen hack)
This is the 16:9 counterpart of the existing widescreen hack, which
is designed for 4:3-native games. This one is designed for games
that natively support 16:9 and are configured to run this way,
or games that have 16:9 widescreen cheats available.

When targeting ultrawide aspect ratios like 21:9, adjusting from 16:9
results in fewer culling issues compared to adjusting from 4:3.
The downside is that in some games, HUD elements will be stretched
up to 16:9 (this depends on the game's widescreen implementation).

You can usually get good results with aspect ratios up to 21:9,
although wider aspect ratios will start exhibiting more and more
culling issues inherent to each game.

The current setting is preserved in existing configurations;
"Try to adjust game to fit" was renamed to "Try to adjust 4:3 game to fit"
for clarity.